Remove BadPowerSpawnChance
All bad powers whose cost is lower than the absolute value of the current karma amount will be valid for backlash, and random selection will be weighted by cost to prioritize powers with a closer cost to the negative karma balance. Did that make any sense? It's 1:00 AM and I can't tell anymore